+++ src/main/resources/db/migration/V2__Create_question_table.sql
@@ -1,10 +1,10 @@
 CREATE TABLE `question` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` bigint(11) NOT NULL AUTO_INCREMENT,
   `title` varchar(50) COLLATE utf8_unicode_ci DEFAULT NULL,
   `description` text COLLATE utf8_unicode_ci,
   `gmt_create` bigint(20) DEFAULT NULL,
   `gmt_modified` bigint(20) DEFAULT NULL,
-  `creator` int(11) DEFAULT NULL,
+  `creator` bigint(11) DEFAULT NULL,
   `comment_count` int(11) NOT NULL DEFAULT '0',
   `view_count` int(11) NOT NULL DEFAULT '0',
   `like_count` int(11) NOT NULL DEFAULT '0',
